update latency logging

+++ src/solution.c
@@ -1230,7 +1230,7 @@ static void solution_thread(void *arg)
       u64 tc_latency = final_tc - current_tc;
       gps_time_t final_gps_time = napcount2gpstime(final_tc);
       double GPS_latency = gpsdifftime(&final_gps_time,&lgf.position_solution.time);
-      log_warn("TOW: %u,  GPS Latency %f, Tick Count Latency %f %u, Processing Time %f %u", pos_llh.tow, GPS_latency, tc_latency, tc_latency * RX_DT_NOMINAL, delta_tc, delta_tc * RX_DT_NOMINAL);
+      log_warn("TOW: %u,  GPS Latency %f, Tick Count Latency %u %f, Processing Time %f %f", pos_llh.tow, GPS_latency, tc_latency, (double)tc_latency * RX_DT_NOMINAL, delta_tc, (double)delta_tc * RX_DT_NOMINAL);
     }
 
     last_spp = chVTGetSystemTime();
